Tweak an argument type of a static function.
* subversion/libsvn_subr/sorts.c
  (bsearch_lower_bound): Change the type of the ELT_SIZE argument from
    'size_t' to 'int', because that's what the only caller wants to pass and
    also because I just committed a doc-string saying its max value is
    INT_MAX :-)


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@873367 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_subr/sorts.c
@@ -167,7 +167,7 @@ static int
 bsearch_lower_bound(const void *key,
                     const void *base,
                     int nelts,
-                    size_t elt_size,
+                    int elt_size,
                     int (*compare_func)(const void *, const void *))
 {
   int lower = 0;
